changed: postinstall-script

+++ postinstall-script.js
@@ -8,11 +8,13 @@
 'use strict';
 
 const fs = require("fs");
+const path = require("path");
 
 // User's local directory
 const userPath = process.env.INIT_CWD
 
-if(__dirname.includes('node_modules')) {
+if(__dirname.includes('node_modules')
+  && fs.existsSync(path.join(userPath, 'config.js')) === false) {
   // Moving files to user's local directory
   fs.cpSync('start', userPath, {
     recursive: true
